@@ -4,8 +4,9 @@
   "description": "API for DataPorts",
   "main": "dist/index.js",
   "scripts": {
-    "build": "npm run clean && babel src -d dist --source-maps",
-    "start": "nodemon --inspect src/index.js --config  nodemon.json --exec babel-node",
+    "build": "npm run clean && babel src -d dist --source-maps --copy-files",
+    "start": "nodemon --inspect src/index.js --config  nodemon.json --exec ./node_modules/.bin/babel-node",
+    "start_prod": "nodemon index.js --config  nodemon.json",
     "serve": "npm run clean && npm run build && cross-env NODE_ENV=production node --inspect=0.0.0.0 dist/index.js",
     "clean": "rimraf dist logs",
     "lint": "eslint 'src/**/*.js' --fix",
@@ -18,38 +19,40 @@
   "author": "JClemente - ProDevelop S.L",
   "license": "ISC",
   "devDependencies": {
-    "@babel/cli": "^7.7.4",
-    "@babel/core": "^7.7.4",
-    "@babel/node": "^7.7.4",
-    "@babel/preset-env": "^7.7.4",
-    "cross-env": "^6.0.3",
-    "eslint": "^6.7.1",
-    "eslint-config-airbnb-base": "^14.0.0",
-    "eslint-config-prettier": "^6.1.0",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-prettier": "^3.1.0",
-    "eslint-plugin-promise": "^4.2.1",
-    "nodemon": "^2.0.1",
-    "prettier": "^1.18.2",
-    "prettier-eslint-cli": "^5.0.0",
-    "rimraf": "^3.0.0"
+    "@babel/cli": "^7.20.7",
+    "@babel/core": "^7.20.12",
+    "@babel/node": "^7.20.7",
+    "@babel/preset-env": "^7.20.2",
+    "eslint": "^8.32.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-promise": "^6.1.1",
+    "prettier": "^2.8.3",
+    "prettier-eslint-cli": "^7.1.0"
   },
   "dependencies": {
-    "adm-zip": "^0.5.5",
-    "axios": "^0.19.2",
-    "axios-cache-adapter": "^2.5.0",
-    "body-parser": "^1.19.0",
+    "adm-zip": "^0.5.10",
+    "axios": "^0.21.1",
+    "axios-cache-adapter": "^2.7.3",
+    "body-parser": "^1.20.1",
     "cors": "^2.8.5",
-    "dotenv": "^8.1.0",
-    "express": "^4.17.1",
-    "express-validator": "^6.3.0",
+    "cross-env": "^7.0.3",
+    "dotenv": "^16.0.3",
+    "express": "^4.18.2",
+    "express-prom-bundle": "^6.6.0",
+    "express-validator": "^6.14.2",
     "express-ws": "^5.0.2",
     "express-zip": "^3.0.0",
-    "mongoose": "^5.12.2",
-    "socket.io": "^4.1.3",
-    "swagger-ui-express": "^4.1.6",
-    "winston": "^3.2.1",
-    "ws": "^7.5.3",
-    "yaml": "^1.10.0"
+    "mongoose": "^6.8.4",
+    "nodemon": "^2.0.20",
+    "rimraf": "^4.1.1",
+    "socket.io": "^4.5.4",
+    "swagger-stats": "^0.99.5",
+    "swagger-ui-express": "^4.6.0",
+    "winston": "^3.8.2",
+    "ws": "^8.12.0",
+    "yaml": "^2.2.1"
   }
 }
