@@ -1,22 +1,14 @@
 const config = {};
-/* const fs = require('fs'); 
-var ADMIN_SECRET_FILE = null
-try {
- ADMIN_SECRET_FILE = fs.readFileSync(process.env.ADMIN_SECRET_FILE, {encoding:'utf8', flag:'r'}); 
-} catch (error) {
-  console.warn("ENV ADMIN_SECRET_FILE not found, getting ADMIN_SECRET_STRING or default value now...")
-}
-*/
+
 config.LOGS_DIR = process.env.LOGS_DIR || 'logs';
 config.PORT = process.env.PORT || 3000;
 
 /**
  * MongoDB
  */
-
 config.mongodb = {};
-config.mongodb.URL = process.env.MONGO_URL || 'mongodb://localhost:27017/dataports-db'; // url para local
-// config.mongodb.URL = process.env.MONGO_URL || 'mongodb://mongo/dataports-db'; // url para desplegar con docker-compose
+config.mongodb.URL =
+  process.env.MONGO_URL || 'mongodb://localhost:27017/dataports-db'; // url para local
 
 /**
  * General Configuration
