@@ -1,156 +1,144 @@
 import express from 'express';
-import { check } from 'express-validator';
+import { authUser, isAuthorized } from '../middlewares/auth';
 import logger from '../config/winston';
-import dataModelService from '../services/dataModel.service';
-
-import util from '../middlewares/util';
-
-import variables from '../utils/variables';
+import * as dataModelService from '../services/dataModel.service';
 import CustomError from '../utils/CustomError';
 
 // eslint-disable-next-line
 import { constructDataModelObj, updateDataModelObj } from '../utils/Mapping';
 
-const router = express.Router();
+// eslint-disable-next-line import/prefer-default-export
+export const router = express.Router();
 
 // routes
-router.get('/', getAll);
-router.post('/', createDataModel_Assigment);
-router.delete('/:id', deleteDataModel);
-router.put('/:id', updateDataModel);
+router.get('/', authUser, isAuthorized, getAll);
+router.post('/', authUser, isAuthorized, createDataModelAssigment);
+router.delete('/:id', authUser, isAuthorized, deleteDataModel);
+router.put('/:id', authUser, isAuthorized, updateDataModel);
 router.post('/getSchema', getJsonSchema);
 
-export default router;
-
 // Implementación
-async function getAll(req, res) {
-	const { query } = req;
-	try {
-		const data = await dataModelService.getAll();
-		res.status(200).json({
-			status: 'OK',
-			message: data
-		});
-	} catch (error) {
-		logger.error(error);
-		if (error instanceof CustomError) {
-			const { message, name, stack, type } = error;
-			return res.status(type).json({
-				status: name,
-				message: message.message
-			});
-		}
-		return res.status(404).json({
-			status: 'Error',
-			message: 'An error ocurred'
-		});
-	}
-	return null;
+async function getAll(_req, res) {
+  try {
+    const data = await dataModelService.getAll();
+    return res.status(200).json({
+      status: 'OK',
+      message: data,
+    });
+  } catch (error) {
+    logger.error(error.toString());
+    if (error instanceof CustomError) {
+      const { message, name, type } = error;
+      return res.status(type).json({
+        status: name,
+        message: message.message,
+      });
+    }
+    return res.status(404).json({
+      status: 'Error',
+      message: 'An error ocurred',
+    });
+  }
 }
 
-async function createDataModel_Assigment(req, res) {
-	const dataModel = constructDataModelObj(req.body);
+async function createDataModelAssigment(req, res) {
+  const dataModel = constructDataModelObj(req.body);
 
-	try {
-		const data = await dataModelService.createDataModel_Assigment(dataModel);
-		res.status(200).json({
-			status: 'OK',
-			message: data
-		});
-	} catch (error) {
-		logger.error(error);
-		if (error instanceof CustomError) {
-			const { message, name, stack, type } = error;
-			return res.status(type).json({
-				status: name,
-				message: message.message
-			});
-		}
-		return res.status(404).json({
-			status: 'Error',
-			message: 'An error occurred'
-		});
-	}
-	return null;
+  try {
+    const data = await dataModelService.createDataModelAssigment(dataModel);
+    return res.status(200).json({
+      status: 'OK',
+      message: data,
+    });
+  } catch (error) {
+    logger.error(error.toString());
+    if (error instanceof CustomError) {
+      const { message, name, type } = error;
+      return res.status(type).json({
+        status: name,
+        message: message.message,
+      });
+    }
+    return res.status(404).json({
+      status: 'Error',
+      message: 'An error occurred',
+    });
+  }
 }
 
 async function deleteDataModel(req, res) {
-	const { params: { id } = { id: null } } = req;
+  const { params: { id } = { id: null } } = req;
 
-	try {
-		const data = await dataModelService.deleteDataModel(id);
-		res.status(200).json({
-			status: 'OK',
-			message: data
-		});
-	} catch (error) {
-		logger.error(error);
-		if (error instanceof CustomError) {
-			const { message, name, stack, type } = error;
-			return res.status(type).json({
-				status: name,
-				message: message.message
-			});
-		}
-		return res.status(404).json({
-			status: 'Error',
-			message: 'An error ocurred'
-		});
-	}
-	return null;
+  try {
+    const data = dataModelService.deleteDataModel(id);
+    return res.status(200).json({
+      status: 'OK',
+      message: data,
+    });
+  } catch (error) {
+    logger.error(error.toString());
+    if (error instanceof CustomError) {
+      const { message, name, type } = error;
+      return res.status(type).json({
+        status: name,
+        message: message.message,
+      });
+    }
+    return res.status(404).json({
+      status: 'Error',
+      message: 'An error ocurred',
+    });
+  }
 }
 
 async function updateDataModel(req, res) {
-	console.log('UPDATE DATA MODEL');
-	const { params: { id } = { id: null } } = req;
-	const dataModel = updateDataModelObj(req.body);
+  logger.debug('UPDATE DATA MODEL');
+  const { params: { id } = { id: null } } = req;
+  const dataModel = updateDataModelObj(req.body);
 
-	try {
-		const data = await dataModelService.updateDataModel(id, dataModel);
-		res.status(200).json({
-			status: 'OK',
-			message: data
-		});
-	} catch (error) {
-		logger.error(error);
-		if (error instanceof CustomError) {
-			const { message, name, stack, type } = error;
-			return res.status(type).json({
-				status: name,
-				message: message.message
-			});
-		}
-		return res.status(404).json({
-			status: 'Error',
-			message: 'An error occurred'
-		});
-	}
-	return null;
+  try {
+    const data = dataModelService.updateDataModel(id, dataModel);
+    return res.status(200).json({
+      status: 'OK',
+      message: data,
+    });
+  } catch (error) {
+    logger.error(error.toString());
+    if (error instanceof CustomError) {
+      const { message, name, type } = error;
+      return res.status(type).json({
+        status: name,
+        message: message.message,
+      });
+    }
+    return res.status(404).json({
+      status: 'Error',
+      message: 'An error occurred',
+    });
+  }
 }
 
 async function getJsonSchema(req, res) {
-	const properties = req.body;
+  const properties = req.body;
 
-	try {
-		console.log(properties);
-		const data = await dataModelService.getJsonSchema(properties);
-		res.status(200).json({
-			status: 'OK',
-			message: data
-		});
-	} catch (error) {
-		logger.error(error);
-		if (error instanceof CustomError) {
-			const { message, name, stack, type } = error;
-			return res.status(type).json({
-				status: name,
-				message: message.message
-			});
-		}
-		return res.status(404).json({
-			status: 'Error',
-			message: 'An error occurred'
-		});
-	}
-	return null;
+  try {
+    const data = await dataModelService.getJsonSchema(properties);
+    return res.status(200).json({
+      status: 'OK',
+      message: data,
+    });
+  } catch (error) {
+    logger.error(error.toString());
+    if (error instanceof CustomError) {
+      const { message, name, type } = error;
+      return res.status(type).json({
+        status: name,
+        message: message.message,
+      });
+    }
+    return res.status(404).json({
+      status: 'Error',
+      message: 'An error occurred',
+    });
+  }
 }
