@@ -1,208 +1,184 @@
 import express from 'express';
 import logger from '../config/winston';
+import { authUser, isAuthorized } from '../middlewares/auth';
 
-import notificationService from '../services/notification.service';
+import * as notificationService from '../services/notification.service';
 import CustomError from '../utils/CustomError';
 
 import { constructNotification } from '../utils/Mapping';
 
 import variables from '../utils/variables';
-import agentService from '../services/agents.service';
-import orionService from '../services/orion.service';
+import * as agentService from '../services/agents.service';
+import { WsSocket } from '../utils/wsSocket';
 
-import wsSocket from '../utils/wsSocket';
+const webSocket = new WsSocket(3010);
 
-import { sleep } from '../utils/utilities';
+// eslint-disable-next-line import/prefer-default-export
+export const router = express.Router();
 
-const webSocket = new wsSocket(3010);
-
-/* const wss = new WebSocket.Server({ port: 8080 });
-wss.on('connection', ws => {
-	ws.on('message', message => {
-		console.log(`Received message => ${message}`);
-	});
-	ws.send('Hello! Message From Server!!');
-}); */
-
-const router = express.Router();
-
-// routes
-router.get('/:id', getOne);
-router.get('/', getAll);
+// secured routes
+router.get('/:id', authUser, isAuthorized, getOne);
+router.get('/', authUser, isAuthorized, getAll);
+router.delete('/:id', authUser, isAuthorized, deleteNotification);
+// required by agents
 router.post('/', create);
-router.delete('/:id', deleteNotification);
 router.patch('/manageOnDemandContainers/:id', manageOnDemandContainers);
-
-export default router;
 
 // Implementation
 async function getOne(req, res) {
-	const { params: { id } = { id: null } } = req;
+  const { params: { id } = { id: null } } = req;
 
-	try {
-		const data = await notificationService.getOne(id);
-		res.status(200).json({
-			status: 'OK',
-			message: data
-		});
-	} catch (error) {
-		logger.error(error);
-		if (error instanceof CustomError) {
-			const { message, name, stack, type } = error;
-			return res.status(type).json({
-				status: name,
-				message: message.message
-			});
-		}
-		return res.status(404).json({
-			status: 'Error',
-			message: 'An error occurred'
-		});
-	}
-	return null;
+  try {
+    const data = await notificationService.getOne(id);
+    return res.status(200).json({
+      status: 'OK',
+      message: data,
+    });
+  } catch (error) {
+    logger.error(error.toString());
+    if (error instanceof CustomError) {
+      const { message, name, type } = error;
+      return res.status(type).json({
+        status: name,
+        message: message.message,
+      });
+    }
+    return res.status(404).json({
+      status: 'Error',
+      message: 'An error occurred',
+    });
+  }
 }
 
 async function getAll(req, res) {
-	const { query } = req;
+  const { limit, offset, ...query } = req.query;
 
-	try {
-		const data = await notificationService.getAll(query);
-		res.status(200).json({
-			status: 'OK',
-			message: data
-		});
-	} catch (error) {
-		logger.error(error);
-		if (error instanceof CustomError) {
-			const { message, name, stack, type } = error;
-			return res.status(type).json({
-				status: name,
-				message: message.message
-			});
-		}
-		return res.status(404).json({
-			status: 'Error',
-			message: 'An error occurred'
-		});
-	}
-	return null;
+  try {
+    const { count, data } = await notificationService.getAll(
+      query,
+      limit,
+      offset
+    );
+    return res.set('x-total-count', count).status(200).json({
+      status: 'OK',
+      message: data,
+      total: count,
+    });
+  } catch (error) {
+    logger.error(error.toString());
+    if (error instanceof CustomError) {
+      const { message, name, type } = error;
+      return res.status(type).json({
+        status: name,
+        message: message.message,
+      });
+    }
+    return res.status(404).json({
+      status: 'Error',
+      message: 'An error occurred',
+    });
+  }
 }
 
 async function deleteNotification(req, res) {
-	const { params: { id } = { id: null } } = req;
-	const query = req.body;
+  const { params: { id } = { id: null } } = req;
+  const query = req.body;
 
-	try {
-		const data = await notificationService.deleteNotification(id, query);
-		res.status(200).json({
-			status: 'OK',
-			message: data
-		});
-	} catch (error) {
-		logger.error(error);
-		if (error instanceof CustomError) {
-			const { message, name, stack, type } = error;
-			return res.status(type).json({
-				status: name,
-				message: message.message
-			});
-		}
-		return res.status(404).json({
-			status: 'Error',
-			message: 'An error occurred'
-		});
-	}
-	return null;
+  try {
+    const data = await notificationService.deleteNotification(id, query);
+    return res.status(200).json({
+      status: 'OK',
+      message: data,
+    });
+  } catch (error) {
+    logger.error(error.toString());
+    if (error instanceof CustomError) {
+      const { message, name, type } = error;
+      return res.status(type).json({
+        status: name,
+        message: message.message,
+      });
+    }
+    return res.status(404).json({
+      status: 'Error',
+      message: 'An error occurred',
+    });
+  }
 }
 
 async function create(req, res) {
-	const query = constructNotification(req.body);
+  const query = constructNotification(req.body);
 
-	try {
-		const data = await notificationService.create(query);
+  try {
+    const data = await notificationService.create(query);
 
-		res.status(200).json({
-			status: 'OK',
-			message: data
-		});
-		console.log('Voy a enviar la notificación por WebSocket');
-		const { socket } = webSocket;
-		// console.log(data);
-		// console.log(socket);
-		const message = `${data.id},${data.type},${data.message}`;
-		socket.send(message);
-		// sleep(1000);
-	} catch (error) {
-		logger.error(error);
-		if (error instanceof CustomError) {
-			const { message, name, stack, type } = error;
-			return res.status(type).json({
-				status: name,
-				message: message.message
-			});
-		}
-		return res.status(404).json({
-			status: 'Error',
-			message: 'An error occurred'
-		});
-	}
-	return null;
+    const message = `${data.id},${data.type},${data.message}`;
+    webSocket.broadcast(message);
+    return res.status(200).json({
+      status: 'OK',
+      message: data,
+    });
+  } catch (error) {
+    logger.error(error.toString());
+    if (error instanceof CustomError) {
+      const { message, name, type } = error;
+      return res.status(type).json({
+        status: name,
+        message: message.message,
+      });
+    }
+    return res.status(404).json({
+      status: 'Error',
+      message: 'An error occurred',
+    });
+  }
 }
 
 async function manageOnDemandContainers(req, res) {
-	const { params: { id } = { id: null } } = req;
+  const { params: { id } = { id: null } } = req;
 
-	if (!id) {
-		return res.status(variables.INTERNAL_SERVER_ERROR).json({
-			status: 'Validation',
-			message: 'Malformed url'
-		});
-	}
+  if (!id) {
+    return res.status(variables.INTERNAL_SERVER_ERROR).json({
+      status: 'Validation',
+      message: 'Malformed url',
+    });
+  }
 
-	try {
-        const inspectAgent = await agentService.inspectAgent(id);
-		
-		let imageID = inspectAgent.Id;
-		let imageName = inspectAgent.Image;
+  try {
+    await agentService.stopAgent(id);
 
-		const stopped = await agentService.stopAgent(id);
-		
-		const removed = await agentService.deleteAgent(id);
+    await agentService.deleteAgent(id);
 
-		const notificationData = {
-			'id': id,
-			'type': 'SUCCESS',
-			'message': 'On-Demand Agent removed',
-			'register': ''
-		}
+    const notificationData = {
+      id,
+      type: 'SUCCESS',
+      message: 'On-Demand Agent removed',
+      register: '',
+    };
 
-		const query = constructNotification(notificationData);
+    const query = constructNotification(notificationData);
 
-		const data_notification = await notificationService.create(query);
+    const dataNotification = await notificationService.create(query);
 
-		res.status(200).json({
-			status: 'OK',
-			message: data_notification
-		});
-		console.log('Voy a enviar una notificación para indicar que borro el contenedor on-demand');
-		const { socket } = webSocket;
-		const message = `${data_notification.id},${data_notification.type},${data_notification.message}`;
-		socket.send(message);
+    const message = `${dataNotification.id},${dataNotification.type},${dataNotification.message}`;
+    webSocket.broadcast(message);
+    return res.status(200).json({
+      status: 'OK',
+      message: dataNotification,
+    });
+  } catch (error) {
+    logger.error(error.toString());
 
-	} catch (error) {
-		logger.error(error);
-
-		if (error instanceof CustomError) {
-			const { message, name, stack, type } = error;
-			return res.status(type).json({
-				status: name,
-				message: message.message
-			});
-		}
-		return res.status(404).json({
-			status: 'Error',
-			message: 'An error occurred'
-		});
-	}
-	return null;
+    if (error instanceof CustomError) {
+      const { message, name, type } = error;
+      return res.status(type).json({
+        status: name,
+        message: message.message,
+      });
+    }
+    return res.status(404).json({
+      status: 'Error',
+      message: 'An error occurred',
+    });
+  }
 }
