@@ -1,12 +1,7 @@
 import { validationResult } from 'express-validator';
 import config from '../config/config';
 
-const middleware = {};
-
-middleware.languageParser = languageParser;
-middleware.sendValidations = sendValidations;
-
-function languageParser(req, res, next) {
+export function languageParser(req, _res, next) {
   const defaultLang = config.const.LANG[0];
   if (req.headers['accept-language'] == null) {
     req.language = defaultLang;
@@ -20,7 +15,7 @@ function languageParser(req, res, next) {
   return null;
 }
 
-function sendValidations(req, res, next) {
+export function sendValidations(req, res, next) {
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
     return res.status(422).json({ errors: errors.array() });
@@ -28,5 +23,3 @@ function sendValidations(req, res, next) {
   next();
   return null;
 }
-
-export default middleware;
