@@ -1,24 +1,24 @@
 import { model, Schema } from 'mongoose';
 
 const mySchema = new Schema(
-	{
-		random_id: {
-			type: 'String'
-		},
-		container_name: {
-			type: 'String'
-		},
-		time_interval: {
-			type: 'Number'
-		},
-		time_unit: {
-			type: 'String'
-		}
-	},
-	{
-		autoCreate: true,
-		timestamps: true
-	}
+  {
+    random_id: {
+      type: 'String',
+    },
+    container_name: {
+      type: 'String',
+    },
+    time_interval: {
+      type: 'Number',
+    },
+    time_unit: {
+      type: 'String',
+    },
+  },
+  {
+    autoCreate: true,
+    timestamps: true,
+  }
 );
 
 mySchema.index({ '$**': 'text' });
