@@ -1,31 +1,27 @@
 import { model, Schema } from 'mongoose';
 
 const mySchema = new Schema(
-	{
-		Hostname: {
-			type: 'String'
-		},
-		Domainname: {
-			type: 'String'
-		},
-		User: {
-			type: 'String'
-		},
-		Entrypoint: {
-			type: 'String'
-		},
-		Image: {
-			type: 'String'
-		},
-		Env: []
-	},
-	{
-		_id: false
-	}
-	/*,
   {
-    timestamps: true
-  }*/
+    Hostname: {
+      type: 'String',
+    },
+    Domainname: {
+      type: 'String',
+    },
+    User: {
+      type: 'String',
+    },
+    Entrypoint: {
+      type: 'String',
+    },
+    Image: {
+      type: 'String',
+    },
+    Env: [],
+  },
+  {
+    _id: false,
+  }
 );
 
 mySchema.index({ '$**': 'text' });
