@@ -1,31 +1,31 @@
 import { model, Schema } from 'mongoose';
 
 const mySchema = new Schema(
-	{
-		source: {
-			type: 'String'
-		},
-		type: {
-			type: 'String'
-		},
-		datamodel: {
-			type: 'String'
-		},
-		constants: {
-			type: 'String'
-		},
-		dockerFile: {
-			type: 'String'
-		},
-		script: {
-			type: 'String'
-		},
-		template: { type: 'Mixed' }
-	},
-	{
-		autoCreate: true,
-		timestamps: true
-	}
+  {
+    source: {
+      type: 'String',
+    },
+    type: {
+      type: 'String',
+    },
+    datamodel: {
+      type: 'String',
+    },
+    constants: {
+      type: 'String',
+    },
+    dockerFile: {
+      type: 'String',
+    },
+    script: {
+      type: 'String',
+    },
+    template: { type: 'Mixed' },
+  },
+  {
+    autoCreate: true,
+    timestamps: true,
+  }
 );
 
 mySchema.index({ '$**': 'text' });
