@@ -1,21 +1,12 @@
 ﻿import { sendInfo } from '../api/cygnus';
 
-const service = {};
-
-service.infoToCygnus = infoToCygnus;
-
-export default service;
+import logger from '../config/winston';
 
 // Implementation
-async function infoToCygnus(entity) {
-  //console.log('Envio a CYGNUS')
-  //console.log(entity)
-	let cygnusEntity = null;
-	try {
-		cygnusEntity = await sendInfo(entity);
-	} catch (error) {
-    //console.log(error);
-		throw new Error(error);
-	}
-	return cygnusEntity;
+// eslint-disable-next-line import/prefer-default-export
+export async function infoToCygnus(entity) {
+  return sendInfo(entity).catch((error) => {
+    logger.error(error.toString());
+    throw new Error(error);
+  });
 }
