@@ -1,90 +1,63 @@
 import PythonTemplate from '../models/PythonTemplate.model';
 
+import logger from '../config/winston';
+
 import { createFilesForZIP } from '../utils/utilities';
 
-const fs = require('fs');
-
-const service = {}
-
-service.getPythonTemplate = getPythonTemplate;
-service.getAll = getAll;
-service.deletePythonTemplate = deletePythonTemplate;
-service.createPythonTemplate = createPythonTemplate;
-service.updatePythonTemplate = updatePythonTemplate;
-service.getZIPTemplate = getZIPTemplate;
-
-export default service
-
 // Implementation
-
-async function getPythonTemplate(id) {
-	let data = [];
-	try {
-	  data = await PythonTemplate.findOne({ _id: id });
-	} catch (error) {
-	  throw new Error(error);
-	}
-	return data;
-}
-  
-async function getAll() {
-	let data = []
-	try {
-	  data = await PythonTemplate.find()
-	} catch (error) {
-	  throw new Error(error)
-	}
-	return data
+export async function getPythonTemplate(id) {
+  return PythonTemplate.findOne({ _id: id })
+    .exec()
+    .catch((error) => {
+      throw new Error(error);
+    });
 }
 
-async function deletePythonTemplate(id) {
-	let data = [];
-	try {
-	  data = await PythonTemplate.deleteOne({ _id: id })
-	} catch (error) {
-	  throw new Error(error);
-	}
-	return data;
+export async function getAll() {
+  return PythonTemplate.find({})
+    .exec()
+    .catch((error) => {
+      throw new Error(error);
+    });
 }
 
-async function createPythonTemplate(query) {
-	let data = null;
-
-	try {
-		data = await new PythonTemplate(query).save();
-	} catch (error) {
-		throw new Error(error);
-	}
-
-	return data;
+export async function deletePythonTemplate(id) {
+  return PythonTemplate.deleteOne({ _id: id })
+    .exec()
+    .catch((error) => {
+      throw new Error(error);
+    });
 }
 
-async function updatePythonTemplate(id, query) {
-	let data = null;
-
-	try {
-		data = await PythonTemplate.findOneAndUpdate({ _id: id}, { $set: query }, { new: true });
-		if (!data) {
-			throw new Error('Not exist');
-		}
-	} catch (error) {
-		throw new Error(error);
-	}
-	return data;
+export async function createPythonTemplate(query) {
+  return new PythonTemplate(query).save().catch((error) => {
+    throw new Error(error);
+  });
 }
 
-function getZIPTemplate(query) {
-	
-	let folderName = '';
+export function updatePythonTemplate(id, query) {
+  let data = null;
 
-	try {
-		console.log(query);
-		
-		folderName = createFilesForZIP(query);
-		
-	} catch (error) {
-		throw new Error(error);
-	}
-	return folderName;
+  try {
+    data = PythonTemplate.findOneAndUpdate(
+      { _id: id },
+      { $set: query },
+      { new: true }
+    ).exec();
+    if (!data) {
+      throw new Error('Not exist');
+    }
+    return data;
+  } catch (error) {
+    throw new Error(error);
+  }
+}
 
+export function getZIPTemplate(query) {
+  try {
+    return createFilesForZIP(query);
+  } catch (error) {
+    logger.error(error.toString());
+    throw new Error(error);
+  }
 }
