@@ -3,19 +3,13 @@
 from datetime import datetime as dt
 from datetime import timezone
 
-from pyngsi.agent import NgsiAgent
-from pyngsi.sources.source import Row, Source
-from pyngsi.sources.source_json import SourceJson
-
-from pyngsi.sink import SinkOrion, SinkStdout
 from pyngsi.ngsi import DataModel
 
 from pyngsi.utils.iso8601 import datetime_to_iso8601
 
 import requests
-import time
 
-from constants import *
+from constants import API_URL, RANDOM_ID, PRIVATE_REPOSITORY, PROJECT_NAME_PRIVATE_REPOSITORY, LINK_PRIVATE_REPOSITORY, URL_PUBLIC_REPOSITORY, CALLBACK_URL
 
 import os
 import sys
@@ -24,224 +18,219 @@ import json
 import jsonschema
 from jsonschema import validate
 
-import base64
+import logging
+import traceback
+import urllib.parse
+logging.basicConfig(level="INFO",
+                    format='%(asctime)s.%(msecs)03d %(levelname)s %(module)s - %(funcName)s: %(message)s',
+                    datefmt='%Y-%m-%d %H:%M:%S',
+                    )
+# TODO: adjust agentName that will be displayed in logs
+logger = logging.getLogger('agentName')
 
-def ValidateJsonSchema(json_data):
+
+def validate_json_schema(json_data):
 
     try:
 
-        isPrivateRepository = PRIVATE_REPOSITORY
+        is_private_repository = PRIVATE_REPOSITORY
 
-        if isPrivateRepository:
+        if is_private_repository:
             # PRIVATE REPOSITORY. GIT
-            print('private repository')
-            is_valid, msg = ValidateJsonSchema_PrivateRepository(json_data)
+            is_valid, msg = validate_json_schema_from_private_repository(
+                json_data)
         else:
             # PUBLIC REPOSITORY. FIWARE
-            is_valid, msg = ValidateJsonSchema_PublicRepository(json_data)
+            is_valid, msg = validate_json_schema_from_public_repository(
+                json_data)
+        return is_valid, msg
 
     except Exception as ex:
-        print('ERROR: ',ex.args)
-        sendNotification('ERROR', ex.args, '')
+        error = f'An error ocurred while validating the schema'
+        logger.exception(error)
+        send_notification('ERROR', error, json.dumps(
+            {"error": str(ex), "stack": traceback.format_exc()}))
 
-    print('IS VALID', is_valid)
-    return is_valid, msg
 
-def ValidateJsonSchema_PrivateRepository(json_data):
+def validate_json_schema_from_private_repository(json_data):
 
     try:
 
-        #BUILD JSON OBJECT TO OBTAIN JSON SCHEMA
+        # BUILD JSON OBJECT TO OBTAIN JSON SCHEMA
         data = {
-          'projectName': PROJECT_NAME_PRIVATE_REPOSITORY,
-          'link': LINK_PRIVATE_REPOSITORY
+            'projectName': PROJECT_NAME_PRIVATE_REPOSITORY,
+            'link': LINK_PRIVATE_REPOSITORY
         }
 
-        #Next line is for Windows deployment
-        #query = requests.post("http://host.docker.internal:3000/dataModel/getSchema", data = data)
+        json_schema = requests.post(
+            f"{API_URL}/dataModel/getSchema", data=data)
 
-        #Next line is for Linux deployment
-        jsonSchema = requests.post("http://172.17.0.1:3000/dataModel/getSchema", data = data)
-
-        validate(instance=json_data, schema=jsonSchema)
+        validate(instance=json_data, schema=json_schema)
 
     except jsonschema.exceptions.SchemaError as e:
-        print(e)
-        print('There is an error with the schema')
+        error = f'here is an error with the schema'
+        logger.exception(error)
     except jsonschema.exceptions.ValidationError as err:
-        print(err)
-        #print('----------------')
-        #print(err.absolute_path)
-        #print('----------------')
-        #print(err.absolute_schema_path)
-        errmsg = "Given JSON data is invalid." + str(err)
+        error = "Given JSON data is invalid." + str(err)
+        logger.exception(error)
         return False, err
     except Exception as ex:
-        print('ERROR: ',ex.args)
-        sendNotification('ERROR', ex.args, '')
+        error = f'An error ocurred while validating the schema'
+        logger.exception(error)
+        send_notification('ERROR', error, json.dumps(
+            {"error": str(ex), "stack": traceback.format_exc()}))
 
     message = "Given JSON data is valid"
     return True, message
 
-def ValidateJsonSchema_PublicRepository(json_data):
+
+def validate_json_schema_from_public_repository(json_data):
 
     try:
 
         schema_not_valid = True
-        urlPublicRepository = URL_PUBLIC_REPOSITORY
+        url_public_repository = URL_PUBLIC_REPOSITORY
 
-        jsonSchema = requests.get(urlPublicRepository).json()
-        #schemaObtained = json.load(jsonSchema)
+        json_schema = requests.get(url_public_repository).json()
 
-        #validate(instance=json_data, schema=jsonSchema)
-        validator = jsonschema.Draft7Validator(jsonSchema)
-        errors = validator.iter_errors(json_data) #get all validation errors
-        print('--------------------------------------------------------------------------------------------------------------')
-        print('TOTAL DE ERRORES', errors)
-        print('--------------------------------------------------------------------------------------------------------------')
+        validator = jsonschema.Draft7Validator(json_schema)
+        errors = validator.iter_errors(json_data)  # get all validation errors
         for error in errors:
-          schema_not_valid = False
-          print('-------')
-          print(error)
-          print('-------')
+            schema_not_valid = False
+            logger.error(error)
 
-    except jsonschema.exceptions.SchemaError as e:
-        print(e)
-        print('There is an error with the schema')
+    except jsonschema.exceptions.SchemaError:
+        error = f'here is an error with the schema'
+        logger.exception(error)
     except jsonschema.exceptions.ValidationError as err:
-        print(err)
-        #print('----------------')
-        #print(err.absolute_path)
-        #print('----------------')
-        #print(err.absolute_schema_path)
-        errmsg = "Given JSON data is invalid." + str(err)
+        error = "Given JSON data is invalid." + str(err)
+        logger.exception(error)
         return False, err
     except Exception as ex:
-        print('ERROR: ',ex.args)
-        sendNotification('ERROR', ex.args, '')
+        error = f'An error ocurred while validating the schema'
+        logger.exception(error)
+        send_notification('ERROR', error, json.dumps(
+            {"error": str(ex), "stack": traceback.format_exc()}))
 
     message = "Given JSON data is valid"
     return schema_not_valid, message
 
-#Function to convert dateTime to string in format date-time ISO8601
-def convertTo_ISO8601(dateObject):
 
-    year = dateObject.year
-    month = dateObject.month
-    day = dateObject.day
-    hour = dateObject.hour
-    minute = dateObject.minute
-    second = dateObject.second
+def convert_to_iso8601(date_object):
+    """ Function to convert dateTime to string in format date-time ISO8601 """
 
-    date_UTC = dt(year, month, day, hour, minute, second, tzinfo=timezone.utc)
+    year = date_object.year
+    month = date_object.month
+    day = date_object.day
+    hour = date_object.hour
+    minute = date_object.minute
+    second = date_object.second
 
-    return datetime_to_iso8601(date_UTC)
+    date_utc = dt(year, month, day, hour, minute, second, tzinfo=timezone.utc)
 
-#Old implementation before to use Cygnus
+    return datetime_to_iso8601(date_utc)
+
+# Old implementation before to use Cygnus
+
+
 def build_entity(register):
 
     try:
 
-        #property register will contain all the fields returned by the API
+        # property register will contain all the fields returned by the API
         # the developer must analyse the data in order to fulfill the Data Model
 
-        #id of the Data Model must follow this pattern: 'urn:ngsi-ld:DataSource:' and the name of DataModel.
-        m = DataModel(id='urn:ngsi-ld:DataSource:parameter_TypeDataModel', type='parameter_TypeDataModel')
-        #Property dataProvider must be a string witht the provider of the data
+        # id of the Data Model must follow this pattern: 'urn:ngsi-ld:DataSource:' and the name of DataModel.
+        m = DataModel(id='urn:ngsi-ld:DataSource:parameter_TypeDataModel',
+                      type='parameter_TypeDataModel')
+        # Property dataProvider must be a string witht the provider of the data
         m.add("dataProvider", 'parameter_dataProvider')
-        #Add as much property as were necessary following the following syntax
+        # Add as much property as were necessary following the following syntax
         parameter_dataModel
 
-        #is_valid, msg = ValidateJsonSchema(m)
+        return m
 
     except Exception as ex:
-        print('ERROR: ',ex.args)
-        sendNotification('ERROR', ex.args, '')
-    
-    return m
+        error = f'Error building entity'
+        logger.exception(error)
+        send_notification('ERROR', error, json.dumps(
+            {"error": str(ex), "stack": traceback.format_exc(), "data": register}))
+
+    return None
+
 
 def build_entity_cygnus(row):
 
-  try:
-    
-    # Returns a datetime object containing the local date and time
-    dateTimeObj = dt.now()
-    currentTime = dateTimeObj.timestamp()
-
-    #We have to read the row object. Next line is an example
-    #coord = row['coord']
-
-    #next line is an example of id of the data model
-    id_DataModel = 'urn:ngsi-ld:DataSource:parameter_TypeDataModel:' + str(coord['lat']) + ':' + str(coord['lon'])
-
-    #next lines are example of how to work with OpenWeather as third-party API
-    main = row['main']
-    pressure = main['pressure']
-    humidity = main['humidity']
-    temp_min = main['temp_min']
-    temp_max = main['temp_max']
-    
-    data = {
-      "id": id_DataModel,
-      "type": 'parameter_TypeDataModel'
-      #Commented lines are an example of how to fulfill the object that will be sended to Cygnus
-      #"dateObserved": "2016-11-30T07:00:00.00Z",
-      #"pressure": {
-      #    "type": "number",
-      #    "metadata": {},
-      #    "value": pressure
-      #},
-      #"humidity": {
-      #    "type": "number",
-      #    "metadata": {},
-      #    "value": humidity
-      #}
-    }
-
-    #is_valid, msg = ValidateJsonSchema(data)
-    #print('--- Validate JSON Schema ---')
-    #print(is_valid)
-    #print(msg)
-
-  except Exception as ex:
-    print('ERROR: ',ex.args)
-    sendNotification('ERROR', ex.args, '')
-
-  return data
-
-def executeQuery():
-
     try:
 
-        # Example of API URL. Next line it's only an example of how to do it
-        #response = requests.get("https://api.openweathermap.org/data/2.5/weather?id=255274&appid=" + APIKEY_OPEN_WEATHER + "&units=metric").json()
-        response = requests.get("parameter_urlAPI").json()
+        # Returns a datetime object containing the local date and time
+        date_time_obj = dt.now()
+        current_time = date_time_obj.timestamp()
+
+        # We have to read the row object. Next line is an example
+        # coord = row['coord']
+
+        # next line is an example of id of the data model
+        id_data_model = 'urn:ngsi-ld:parameter_TypeDataModel:' + \
+            str(coord['lat']) + ':' + str(coord['lon'])
+
+        # next lines are example of how to work with OpenWeather as third-party API
+        main = row['main']
+        pressure = main['pressure']
+        humidity = main['humidity']
+        temp_min = main['temp_min']
+        temp_max = main['temp_max']
+
+        return {
+            "id": id_data_model,
+            "type": 'parameter_TypeDataModel'
+            # Commented lines are an example of how to fulfill the object that will be sended to Cygnus
+            # "dateObserved": "2016-11-30T07:00:00.00Z",
+            # "pressure": {
+            #    "type": "number",
+            #    "metadata": {},
+            #    "value": pressure
+            # },
+            # "humidity": {
+            #    "type": "number",
+            #    "metadata": {},
+            #    "value": humidity
+            # }
+        }
 
     except Exception as ex:
-        print('ERROR: ',ex.args)
-        sendNotification('ERROR', ex.args, '')
+        error = f'Error building entity'
+        logger.exception(error)
+        send_notification('ERROR', error, json.dumps(
+            {"error": str(ex), "stack": traceback.format_exc(), "data": row}))
 
-    return response
+    return None
 
-def sendNotification(notificationType, messageText, messageSended):
+
+def execute_query():
+    """ Method that requests data to an external URL and returns it. If the call fails, None is returned"""
+
+    # Example of API URL. Next line it's only an example of how to do it
+    # response = requests.get("https://api.openweathermap.org/data/2.5/weather?id=255274&appid=" + APIKEY_OPEN_WEATHER + "&units=metric").json()
+    response = requests.get("parameter_urlAPI")
+    if response.text != None:
+        decoded_data = response.text.encode().decode('utf-8-sig')
+        return json.loads(decoded_data)
+    else:
+        return None
+
+
+def send_notification(notification_type, message_text, message_sent):
 
     try:
+        # RANDOM_ID is a unique identifier that allow notifications to be linked to the container they belong to
+        random_id = urllib.parse.quote(
+            os.getenv("RANDOM_ID", RANDOM_ID), safe='')
 
-        #print('Method that sends notification to MongoDB. Could be successful or an error.')
+        query_info = requests.get(f"{API_URL}/info/{random_id}")
 
-        #RANDOM_ID is a unique identifier that allow notifications to be linked to the container they belong to
-        random_ID = os.getenv("RANDOM_ID", RANDOM_ID)
-        #print('random_ID', random_ID)
-        # IP --> 127.0.0.1 out of the platform (debug time)
-        #query_info = requests.get("http://127.0.0.1:3000/info/" + random_ID)
+        logger.debug(f'answer from /info: {query_info.text}')
 
-        #Next line is for Windows deployment
-        #query_info = requests.get("http://host.docker.internal:3000/info/" + random_ID)
-
-        #Next line is for Linux deployment
-        query_info = requests.get("http://172.17.0.1:3000/info/" + random_ID)
-
-        #print('answer', query_info.text)
         response_dict = json.loads(query_info.text)
         message = response_dict['message']
 
@@ -249,118 +238,135 @@ def sendNotification(notificationType, messageText, messageSended):
 
         data = {
             "id": container_name,
-            "type": notificationType,
-            "message": messageText,
-            "register": messageSended
+            "type": notification_type,
+            "message": message_text,
+            "register": message_sent
         }
-        
-        #Next line is for Windows deployment
-        #query = requests.post("http://host.docker.internal:3000/notification", data = data)
+        # Next line is for Windows deployment
+        query = requests.post(f'{API_URL}/notification', data=data)
 
-        #Next line is for Linux deployment
-        query = requests.post("http://172.17.0.1:3000/notification", data = data)
-
-        #extracting response text
+        # extracting response text
         response = query.text
-        print('response: ', response)
+        logger.debug(f'response from /notification: {response}')
+        return container_name
+    except Exception:
+        logger.exception(
+            f'Could not send notification with payload:\ntype="{notification_type}"\nmessage="{message_text}"\ndata={message_sent}"')
 
-    except Exception as ex:
-        print('ERROR: ',ex.args)
-    
-    return container_name
-    
+
 def main():
+    """ Method that runs the import logic"""
 
     try:
 
-        container_name = ''
+        send_notification(
+            'SUCCESS', 'Agent execution begins.', '')
 
-        container_name = sendNotification('SUCCESS', 'Agent execution begins.', '')
+        array_json = []
 
-        arrayJson = []
+        head = {"Content-Type": "application/json",
+                "Accept": "application/json"}
 
-        head = {"Content-Type": "application/json", "Accept": "application/json"}
+        response = execute_query()
+        data_model = build_entity_cygnus(response)
 
-        response = executeQuery()
-        print(response)
-        #dataModel = build_entity(response)
-        dataModel = build_entity_cygnus(response)
-        #append dataModel object to the arrayJson variable
-        arrayJson.append(dataModel)
-        
-        #SEND DATA TO CYGNUS URL IN JSON FORMAT
-        print('-----------------------------------------------------------')
-        print('data', arrayJson)
-        print('-----------------------------------------------------------')
-        #Next line is for Windows deployment
-        #query_result = requests.post('http://host.docker.internal:3000/cygnus', json= arrayJson, headers=head)
+        if data_model != None:
+            total_count = 0
+            counter = 0
 
-        #Next line is for Linux deployment
-        query_result = requests.post('http://172.17.0.1:3000/cygnus', json= arrayJson, headers=head)
-        print('---- Envio ----', query_result.text)
+            # === on-demand
+            array_json = []
+            for item in data_model:
+                entity = build_entity_cygnus(item)
+                if entity == None:
+                    # None indicates that the registry could not be mapped
+                    continue
+                array_json.append(entity)
+                counter += 1
+                total_count += 1
 
-        #-------- SEND INFORMATION ABOUT THE COLLECTION ON CYGNUS --------
-        cygnus_id = dataModel['id']
-        cygnus_type = dataModel['type']
+                # fargment the data in batches of N elements when sending to Cygnus
+                if counter >= 25:
+                    requests.post(
+                        f'{API_URL}/cygnus', json=array_json, headers=head)
+                    logger.debug('Block of 25 messages sent')
+                    counter = 0
+                    array_json = []
+
+            if counter != 0:
+                # -- FOR SENDING THE LAST BLOCK --
+                response = requests.post(
+                    f'{API_URL}/cygnus', json=array_json, headers=head)
+
+                logger.debug(f'Block of {len(array_json)} messages sent.')
+
+        # append dataModel object to the arrayJson variable
+        array_json.append(data_model)
+
+        # Next line is for Linux deployment
+        requests.post(
+            f'{API_URL}/cygnus', json=array_json, headers=head)
+
+        # -------- SEND INFORMATION ABOUT THE COLLECTION ON CYGNUS --------
+        cygnus_id = data_model['id']
+        cygnus_type = data_model['type']
         name_cygnus_collection = 'col_/_' + cygnus_id + '_' + cygnus_type
-        data_Cygnus = {
+        data_cygnus = {
             "id": cygnus_id,
             "type": cygnus_type,
             "cygnus_collection": name_cygnus_collection,
             "cygnus_database": "db_default",
-            "total": 1
+            "total": total_count
         }
 
-        #INFO about cygnus for ICCS API
-        #Next line is for Windows deployment
-        #query_infoToCygnus = requests.post('http://host.docker.internal:3000/cygnusInformation', data = data_Cygnus)
+        # INFO about cygnus for ICCS API
+        query_info_to_cygnus = requests.post(
+            f'{API_URL}/cygnusInformation', data=data_cygnus)
+        send_notification(
+            'SUCCESS', f'Imported {len(array_json)} records', '')
+        logger.debug(
+            f'Data send to cygnus returned {query_info_to_cygnus.status_code} and answered {query_info_to_cygnus.text}')
 
-        #Next line is for Linux deployment
-        query_infoToCygnus = requests.post("http://172.17.0.1:3000/cygnusInformation", data = data_Cygnus)
-        print('---- Envio Info To Cygnus ----', query_infoToCygnus.text)
+        send_notification(
+            'SUCCESS', 'Agent execution ends.', array_json)
 
+        stop_container_execution()
 
-        container_name = sendNotification('SUCCESS', 'Agent execution ends.', arrayJson)
-        print('NOMBRE CONTENEDOR', container_name)
-
-        #DELETE THE REGISTER THAT ASSOCIATES RANDOM_ID WITH THE NAME OF THE CONTAINER
-        #RANDOM_ID is a unique identifier that allow notifications to be linked to the container they belong to
-        random_ID = os.getenv("RANDOM_ID", RANDOM_ID)
-        #Next line is for Windows deployment
-        #query_deleteInfo = requests.delete("http://host.docker.internal:3000/info/" + random_ID)
-
-        #Next line is for Linux deployment
-        query_deleteInfo = requests.delete("http://172.17.0.1:3000/info/" + random_ID)
-        print('---- Remove association container to random_id ----', query_deleteInfo.text)
-
-        #Lines responsible for deleting on-demand containers
-        print('--------------------------------------------------------------------------------------------------------------')
-        print("------------------------------------ LET'S MANAGE THE ON-DEMAND CONTAINER ------------------------------------")
-        #Next line is for Windows deployment
-        #query_stop_on_demand = requests.patch("http://host.docker.internal:3000/notification/manageOnDemandContainers/" + container_name)
-
-        #Next line is for Linux deployment
-        query_stop_on_demand = requests.patch("http://172.17.0.1:3000/notification/manageOnDemandContainers/" + container_name)
-        print('---- Contenedor parado ----', query_stop_on_demand.text)
-    
     except Exception as ex:
-        print('ERROR: ',ex.args)
-        #print(type(ex))
-        #print(ex)
-        sendNotification('ERROR', ex.args, '')
+        error = f'An error ocurred while processing importing data'
+        logger.exception(error)
+        send_notification('ERROR', error, json.dumps(
+            {"error": str(ex), "stack": traceback.format_exc()}))
+
+
+def stop_container_execution():
+
+    container_name = send_notification('SUCCESS', 'Agent execution ends.', '')
+    logger.info(f'Stopping container {container_name}')
+
+    # DELETE THE REGISTER THAT ASSOCIATES RANDOM_ID WITH THE NAME OF THE CONTAINER
+    # RANDOM_ID is a unique identifier that allow notifications to be linked to the container they belong to
+    random_id = os.getenv("RANDOM_ID", RANDOM_ID)
+
+    # Next line is for Linux deployment
+    query_delete_info = requests.delete(
+        f'{API_URL}/info/{urllib.parse.quote(random_id, safe="")}')
+    logger.info(
+        f'Removed association of container {container_name} to id {random_id}. Response: {query_delete_info.text}')
+
+    logger.info(f'Managing on-demand container {container_name}')
+    query_stop_on_demand = requests.patch(
+        f'{API_URL}/notification/manageOnDemandContainers/{container_name}')
+    logger.info(f'Container stopped with result {query_stop_on_demand.text}')
+
 
 if __name__ == '__main__':
-    lon = len(sys.argv)#Length of input variables
-    
-    #print(lon)
-    print(sys.argv[0])#we start counting from position 0
+    lon = len(sys.argv)  # Length of input variables
 
-    #----- SELECT CALLBACK URL -------
+    # ----- SELECT CALLBACK URL -------
     if lon > 1:
-        #print('lon > 1')
         callback_url = sys.argv[1]
     else:
-        #print('lon <= 1')
         callback_url = os.getenv("CALLBACK_URL", CALLBACK_URL)
 
     main()
