@@ -3,11 +3,7 @@
 from datetime import datetime as dt
 from datetime import timezone
 
-from pyngsi.agent import NgsiAgent
-from pyngsi.sources.source import Row, Source
-from pyngsi.sources.source_json import SourceJson
-
-from pyngsi.sink import SinkOrion, SinkStdout
+from pyngsi.sink import SinkOrion
 from pyngsi.ngsi import DataModel
 
 from pyngsi.utils.iso8601 import datetime_to_iso8601
@@ -15,9 +11,10 @@ from pyngsi.utils.iso8601 import datetime_to_iso8601
 import requests
 import time
 
-from pyngsi.scheduler import Scheduler, UNIT
+import schedule
 
-from constants import *
+from constants import API_URL, BASE_URL, LINK_PRIVATE_REPOSITORY, PRIVATE_REPOSITORY, PROJECT_NAME_PRIVATE_REPOSITORY, \
+    URL_PUBLIC_REPOSITORY, RANDOM_ID, TIME_INTERVAL, TIME_UNIT, switch
 
 import os
 import sys
@@ -26,301 +23,283 @@ import json
 import jsonschema
 from jsonschema import validate
 
-def ValidateJsonSchema(json_data):
+import logging
+import urllib.parse
+import traceback
+logging.basicConfig(level="INFO",
+                    format='%(asctime)s.%(msecs)03d %(levelname)s %(module)s - %(funcName)s: %(message)s',
+                    datefmt='%Y-%m-%d %H:%M:%S',
+                    )
+# TODO: define the agentName that will be displayed in logs
+logger = logging.getLogger(f'agentName')
+sink = SinkOrion(BASE_URL, 1026)
+
+
+def validate_json_schema(json_data):
+    """ Function that validates data against a jsonSchema """
 
     try:
 
-        isPrivateRepository = PRIVATE_REPOSITORY
+        is_private_repository = PRIVATE_REPOSITORY
 
-        if isPrivateRepository:
+        if is_private_repository:
             # PRIVATE REPOSITORY. GIT
-            print('private repository')
-            is_valid, msg = ValidateJsonSchema_PrivateRepository(json_data)
+            is_valid, msg = validate_json_schema_from_private_repository(
+                json_data)
         else:
             # PUBLIC REPOSITORY. FIWARE
-            is_valid, msg = ValidateJsonSchema_PublicRepository(json_data)
+            is_valid, msg = validate_json_schema_from_public_repository(
+                json_data)
+        return is_valid, msg
 
     except Exception as ex:
-        print('ERROR: ',ex.args)
-        sendNotification('ERROR', ex.args, '')
+        error = f'An error ocurred while validating the schema'
+        logger.exception(error)
+        send_notification('ERROR', error, json.dumps(
+            {"error": str(ex), "stack": traceback.format_exc()}))
 
-    print('IS VALID', is_valid)
-    return is_valid, msg
 
-def ValidateJsonSchema_PrivateRepository(json_data):
+def validate_json_schema_from_private_repository(json_data):
 
     try:
 
-        #BUILD JSON OBJECT TO OBTAIN JSON SCHEMA
+        # BUILD JSON OBJECT TO OBTAIN JSON SCHEMA
         data = {
-          'projectName': PROJECT_NAME_PRIVATE_REPOSITORY,
-          'link': LINK_PRIVATE_REPOSITORY
+            'projectName': PROJECT_NAME_PRIVATE_REPOSITORY,
+            'link': LINK_PRIVATE_REPOSITORY
         }
 
-        #Next line is for Windows deployment
-        #query = requests.post("http://host.docker.internal:3000/dataModel/getSchema", data = data)
+        json_schema = requests.post(
+            f"{API_URL}/dataModel/getSchema", data=data)
 
-        #Next line is for Linux deployment
-        jsonSchema = requests.post("http://172.17.0.1:3000/dataModel/getSchema", data = data)
+        validate(instance=json_data, schema=json_schema)
 
-        validate(instance=json_data, schema=jsonSchema)
-
-    except jsonschema.exceptions.SchemaError as e:
-        print(e)
-        print('There is an error with the schema')
+    except jsonschema.exceptions.SchemaError:
+        error = f'here is an error with the schema'
+        logger.exception(error)
     except jsonschema.exceptions.ValidationError as err:
-        print(err)
-        #print('----------------')
-        #print(err.absolute_path)
-        #print('----------------')
-        #print(err.absolute_schema_path)
-        errmsg = "Given JSON data is invalid." + str(err)
+        error = "Given JSON data is invalid." + str(err)
+        logger.exception(error)
         return False, err
     except Exception as ex:
-        print('ERROR: ',ex.args)
-        sendNotification('ERROR', ex.args, '')
+        error = f'An error ocurred while validating the schema'
+        logger.exception(error)
+        send_notification('ERROR', error, json.dumps(
+            {"error": str(ex), "stack": traceback.format_exc()}))
 
     message = "Given JSON data is valid"
     return True, message
 
-def ValidateJsonSchema_PublicRepository(json_data):
+
+def validate_json_schema_from_public_repository(json_data):
 
     try:
 
-        urlPublicRepository = URL_PUBLIC_REPOSITORY
+        schema_not_valid = True
+        url_public_repository = URL_PUBLIC_REPOSITORY
 
-        jsonSchema = requests.get(urlPublicRepository).json()
-        #schemaObtained = json.load(jsonSchema)
+        json_schema = requests.get(url_public_repository).json()
 
-        validate(instance=json_data, schema=jsonSchema)
+        validator = jsonschema.Draft7Validator(json_schema)
+        errors = validator.iter_errors(json_data)  # get all validation errors
+        for error in errors:
+            schema_not_valid = False
+            logger.error(error)
 
-    except jsonschema.exceptions.SchemaError as e:
-        print(e)
-        print('There is an error with the schema')
+    except jsonschema.exceptions.SchemaError:
+        error = f'here is an error with the schema'
+        logger.exception(error)
     except jsonschema.exceptions.ValidationError as err:
-        print(err)
-        #print('----------------')
-        #print(err.absolute_path)
-        #print('----------------')
-        #print(err.absolute_schema_path)
-        errmsg = "Given JSON data is invalid." + str(err)
+        error = "Given JSON data is invalid." + str(err)
+        logger.exception(error)
         return False, err
     except Exception as ex:
-        print('ERROR: ',ex.args)
-        sendNotification('ERROR', ex.args, '')
+        error = f'An error ocurred while validating the schema'
+        logger.exception(error)
+        send_notification('ERROR', error, json.dumps(
+            {"error": str(ex), "stack": traceback.format_exc()}))
 
     message = "Given JSON data is valid"
-    return True, message
+    return schema_not_valid, message
 
-def DataModel_response(register):
 
+def build_entity(item) -> DataModel:
+    """
+    Function that receives an input object and returns a DataModel with mapped fields
+    If an error occurs mapping data, None should be returned
+    """
     try:
 
-        #property register will contain all the fields returned by the API
-        # the developer must analyse the data in order to fulfill the Data Model
-
-        #id of the Data Model must follow this pattern: 'urn:ngsi-ld:DataSource:' and the name of DataModel. Next line is an example
-        m = DataModel(id='urn:ngsi-ld:DataSource:parameter_TypeDataModel', type='parameter_TypeDataModel')
-        #Property dataProvider must be a string witht the provider of the data
+        # TODO: adjust the id pattern here
+        m = DataModel(id='urn:ngsi-ld:parameter_TypeDataModel',
+                      type='parameter_TypeDataModel')
+        # Property dataProvider must be a string witht the provider of the data
         m.add("dataProvider", 'parameter_dataProvider')
-        #Add as much property as were necessary following the following syntax
+        # Add as much property as were necessary following the following syntax
+        # TODO: verify these mappings are correct
         parameter_dataModel
-
-        #is_valid, msg = ValidateJsonSchema(m)
-
-    except Exception as ex:
-        print('ERROR: ',ex.args)
-        sendNotification('ERROR', ex.args, '')
-
-    return m
-
-def build_entity(row: Row) -> DataModel:
-
-    try:
-
-        sendNotification('SUCCESS', 'Agent execution begins.', '')
-
-        response = executeQuery()
-
-        model = DataModel_response(response)
-
-        sendNotification('SUCCESS', 'Agent execution ends.', model.json())
+        return m
 
     except Exception as ex:
-        print('ERROR: ',ex.args)
-        sendNotification('ERROR', ex.args, '')
+        error = f'Failed mapping row with data {json.dumps(item)}'
+        logger.exception(error)
+        send_notification(
+            'ERROR', error, json.dumps({"error": str(ex), "stack": traceback.format_exc(), "row": item}))
 
-    return model
+    return None
 
-def executeQuery():
+
+def execute_query():
+    """
+    Method that calls the external API and returns an array with data
+    """
+
+    # Example of API URL. Next line it's only an example of how to do it
+    # response = requests.get("https://api.openweathermap.org/data/2.5/weather?id=255274&appid=" + APIKEY_OPEN_WEATHER + "&units=metric").json()
+    # TODO: if the expternal API requires query params or special headers, add here
+    response = requests.get("parameter_urlAPI")
+    values = []
+    if response.text != None:
+        decoded_data = response.text.encode().decode('utf-8-sig')
+        # TODO: if endpoint already returns an array, adjust the following line
+        values.append(json.loads(decoded_data))
+    return values
+
+
+def send_notification(notification_type, message_text, message_sent):
+    """
+    Helper function that sends a notification to the DAC.
+    """
 
     try:
+        # RANDOM_ID is a unique identifier that allow notifications to be linked to the container they belong to
+        random_id = urllib.parse.quote(
+            os.getenv("RANDOM_ID", RANDOM_ID), safe='')
 
-        # Example of API URL. Next line it's only an example of how to do it
-        #response = requests.get("https://api.openweathermap.org/data/2.5/weather?id=255274&appid=" + APIKEY_OPEN_WEATHER + "&units=metric").json()
-        response = requests.get("parameter_urlAPI").json()
-    
-    except Exception as ex:
-        print('ERROR: ',ex.args)
-        sendNotification('ERROR', ex.args, '')
+        query_info = requests.get(f"{API_URL}/info/{random_id}")
 
-    return response
+        logger.debug(f'answer from /info: {query_info.text}')
 
-def sendNotification(notificationType, messageText, messageSended):
-
-    try:
-
-        #print('Method that sends notification to MongoDB. Could be successful or an error.')
-
-        #RANDOM_ID is a unique identifier that allow notifications to be linked to the container they belong to
-        random_ID = os.getenv("RANDOM_ID", RANDOM_ID)
-
-        filter = {
-            "random_id": random_ID,
-            "time_interval": time_interval,
-            "time_unit": unit
-        }
-
-        #------------------- STARTS OLD VERSION ---------------------------------
-        #Next line is for Linux deployment
-        #query_info = requests.get("http://172.17.0.1:3000/info/" + random_ID)
-               
-        #Next line is working locally
-        #query_info = requests.get("http://127.0.0.1:3000/info/" + random_ID)
-        
-        #Next line is for Windows deployment
-        #query_info = requests.get("http://host.docker.internal:3000/info/" + random_ID)
-        #print('answer', query_info.text)
-        
-        #------------------- ENDS OLD VERSION ---------------------------------
-
-        #---- new query that links random_id, cotainer and some parameters
-        # Linux Deployment
-        query_info = requests.post("http://172.17.0.1:3000/info/filtered", data = filter)
-        #Next line is working locally
-        #query_info = requests.post("http://127.0.0.1:3000/info/filtered", data = filter)
-        #Next line is for Windows deployment
-        #query_info = requests.post("http://host.docker.internal:3000/info/filtered", data = filter)
-        print('answer: ', query_info.text)
-        
         response_dict = json.loads(query_info.text)
-        #print('----------------------------')
-        #print(response_dict)
-        #print('----------------------------')
-        message = response_dict['message'][0]
-        #print('----------------------------')
-        #print(message)
-        #print('----------------------------')
+        message = response_dict['message']
 
         container_name = message['container_name']
 
         data = {
             "id": container_name,
-            "type": notificationType,
-            "message": messageText,
-            "register": messageSended
+            "type": notification_type,
+            "message": message_text,
+            "register": message_sent
         }
-        
-        #Next line is for Linux deployment
-        query = requests.post("http://172.17.0.1:3000/notification", data = data)
+        # Next line is for Windows deployment
+        query = requests.post(f'{API_URL}/notification', data=data)
 
-        #Next line is for Windows deployment
-        #query = requests.post("http://http://127.0.0.1:3000/notification", data = data)
-
-        #Local Deployment
-        #query = requests.post("http://host.docker.internal:3000/notification", data = data)
-
-        #extracting response text
+        # extracting response text
         response = query.text
-        print('response: ', response)
+        logger.debug(f'response from /notification: {response}')
+        return container_name
+    except Exception:
+        logger.exception(
+            f'Could not send notification with payload:\ntype="{notification_type}"\nmessage="{message_text}"\ndata={message_sent}"')
 
-    except Exception as ex:
-        print('ERROR: ',ex.args)
 
-#Function to convert dateTime to string in format date-time ISO8601
-def convertTo_ISO8601(dateObject):
+def convert_to_iso8601(date_object):
+    """ Function to convert dateTime to string in format date-time ISO8601 """
 
-    year = dateObject.year
-    month = dateObject.month
-    day = dateObject.day
-    hour = dateObject.hour
-    minute = dateObject.minute
-    second = dateObject.second
+    year = date_object.year
+    month = date_object.month
+    day = date_object.day
+    hour = date_object.hour
+    minute = date_object.minute
+    second = date_object.second
 
-    date_UTC = dt(year, month, day, hour, minute, second, tzinfo=timezone.utc)
+    date_utc = dt(year, month, day, hour, minute, second, tzinfo=timezone.utc)
 
-    return datetime_to_iso8601(date_UTC)
+    return datetime_to_iso8601(date_utc)
 
-def main():
 
+def push_orion(entity):
+    """ Method that writes data to Orion broker"""
+    sink.write(entity.json())
+
+
+def create_ps_entities():
+    """ Method that is executed periodically to gather data from the remote API"""
     try:
 
-        print('Process the data')
+        send_notification(
+            'SUCCESS', 'Agent execution begins.', '')
 
-        print('Call the method responsible for make the query to the API')
+        json_response = execute_query()
+        if json_response == None or len(json_response) == 0:
+            logger.warning('No data received')
+        else:
+            logger.info(f'Received {len(json_response)} elements')
 
-        response = executeQuery()
-        #print(response)
-
-        print('Create the source')
-        # provider must be the name of the third-party API or a string to identify the API
-        src = SourceJson(response, provider='parameter_dataProvider')
-
-        # if you want to see the data in development time use the next line
-        sink = SinkStdout()
-        # Change the IP of the ORION and its port
-        #sink = SinkOrion("parameter_urlORION", parameter_orionPORT)
-
-        agent = NgsiAgent.create_agent(src, sink, process=build_entity)
-
-        print('Frequency is:', unit)
-
-        with switch(unit) as s:
-            if s.case('SECONDS', True):
-                scheduler = Scheduler(agent, interval=int(time_interval), unit=UNIT.seconds)
-            if s.case('MINUTES', True):
-                scheduler = Scheduler(agent, interval=int(time_interval), unit=UNIT.minutes)
-            if s.case('HOURS', True):
-                scheduler = Scheduler(agent, interval=int(time_interval), unit=UNIT.hours)
-            if s.case('DAYS', True):
-                scheduler = Scheduler(agent, interval=int(time_interval), unit=UNIT.days)
-            if s.default():
-                scheduler = Scheduler(agent, interval=int(time_interval), unit=UNIT.minutes)
-
-        scheduler.run()
+            sent = 0
+            failed = 0
+            for item in json_response:
+                entity = build_entity(item)
+                if entity is not None:
+                    push_orion(entity)
+                    sent += 1
+                else:
+                    failed += 1
+            send_notification(
+                'SUCCESS', f'Agent execution ended successfully. {sent} registries imported, {failed} registries failed', '')
 
     except Exception as ex:
-        print('ERROR: ',ex.args)
-        #print(type(ex))
-        #print(ex)
-        sendNotification('ERROR', ex.args, '')
+        logger.exception('The agent finished with error')
+        send_notification('ERROR', 'Agent execution failed',
+                          json.dumps({"error": str(ex), "stack": traceback.format_exc()}))
+
+
+def main():
+    try:
+        with switch(unit) as s:
+            if s.case('SECONDS', True):
+                schedule.every(int(time_interval)).seconds.do(
+                    create_ps_entities)
+            if s.case('MINUTES', True):
+                schedule.every(int(time_interval)).minutes.do(
+                    create_ps_entities)
+            if s.case('HOURS', True):
+                schedule.every(int(time_interval)).hours.do(create_ps_entities)
+            if s.case('DAYS', True):
+                schedule.every(int(time_interval)).days.do(create_ps_entities)
+            if s.default():
+                schedule.every(int(time_interval)).days.do(create_ps_entities)
+
+        while True:
+            schedule.run_pending()
+            time.sleep(1)
+
+    except Exception as ex:
+        error = 'The agent failed on its execution'
+        logger.exception(error)
+        send_notification('ERROR', error,
+                          json.dumps({"error": str(ex), "stack": traceback.format_exc()}))
+
 
 if __name__ == '__main__':
-    #---  LOAD THE ENVIRONMENTAL VARIABLES ---
-    
-    lon = len(sys.argv)#Length of input variables
-    
-    #print(lon)
-    #print(sys.argv[0])#we start counting from position 0
+    # ---  LOAD THE ENVIRONMENTAL VARIABLES ---
 
-    #----- SELECT TIME INTERVAL -------
+    lon = len(sys.argv)  # Length of input variables
+
+    # ----- SELECT TIME INTERVAL -------
     if lon > 1:
-        print('lon > 1')
         time_interval = sys.argv[1]
     else:
-        print('lon <= 1')
         time_interval = os.getenv("TIME_INTERVAL", TIME_INTERVAL)
 
-    #----- SELECT TIME UNIT -------
+    # ----- SELECT TIME UNIT -------
     if lon > 2:
-        print('lon > 2')
         unit = sys.argv[2]
     else:
-        print('lon <= 2')
         unit = os.getenv("TIME_UNIT", TIME_UNIT)
 
-    #-- copy the above code if you want to add more environmental variables.
-    #-- these variables must be written with upper case and also in the constants.py. (See TIME_INTERVAL and TIME_UNIT).
-    
-    #---  EXECUTE THE MAIN FUNCTION ---
+    # -- copy the above code if you want to add more environmental variables.
+    # -- these variables must be written with upper case and also in the constants.py. (See TIME_INTERVAL and TIME_UNIT).
+
+    # ---  EXECUTE THE MAIN FUNCTION ---
     main()
